fix(libs): remove plugin SDK cycles
