+++ libs/src/callable.apis.ts
@@ -1,9 +1,11 @@
-import { PluginLocal } from './LSPlugin.core'
-
 /**
  * These APIs run in the host and are callable by users,
  * similar to logseq.apis.cljs but for JS/TS.
  */
+
+type PluginLocal = {
+  sdk?: Record<string, any>
+}
 
 export function setSDKMetadata(this: PluginLocal, data: any) {
   if (this?.sdk && data) {
